use long instead of long long for int64_t (for better conversion

+++ include/dlh/types.hpp
@@ -8,16 +8,16 @@
 typedef unsigned char      uint8_t;
 typedef unsigned short     uint16_t;
 typedef unsigned int       uint32_t;
-typedef unsigned long long uint64_t;
-typedef unsigned long long uintptr_t;
+typedef unsigned long      uint64_t;
+typedef unsigned long      uintptr_t;
 
 typedef __SIZE_TYPE__      size_t;
 
 typedef char               int8_t;
 typedef short              int16_t;
 typedef int                int32_t;
-typedef long long          int64_t;
-typedef long long          intptr_t;
+typedef long               int64_t;
+typedef long               intptr_t;
 
 typedef long int           ssize_t;
